requirements.txt
chroma_hnswlib==0.7.3
langchain==0.1.16
langchain_community==0.0.33
numpy==1.26.4
pandas==1.5.3
psutil==5.9.7
pytest==7.0.1
Requests==2.31.0
scikit_learn==1.2.2
sentence_transformers==2.7.0
streamlit==1.24.0
